Module thor::wallet

Modules

account

committee

delegation

discrimination

utxo

Structs

Block

Block is an element of the blockchain it contains multiple transaction and a reference to the parent block. Alongside with the position of that block in the chain.

Input

Generalized input which have a specific input value, and either contains an account reference or a TransactionSignDataHash+index

LinearFee

Linear fee using the basic affine formula COEFFICIENT * bytes(COUNT(tx.inputs) + COUNT(tx.outputs)) + CONSTANT + CERTIFICATE*COUNT(certificates).

Milli

SpendingCounter

Spending counter associated to an account.

TransactionBindingAuthData

UpdateProposal

UpdateVote

Enums

Fragment

All possible messages recordable in the content

SignedCertificate

Wallet

WalletError

Type Definitions

ConsensusVersion

HeaderId

PoolId

Pool ID

WalletAlias